requirements.txt
gradio
numpy
python-dotenv
langchain-nvidia-ai-endpoints
